# april27finale-proper-data-addition/april27finale-proper-data-addition/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QuickBooks-style business platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceInput"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Invoice deleted
  /invoices/{id}/pay:
    post:
      operationId: payInvoice
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceInput"
      responses:
        "200":
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List all purchase orders
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderInput"
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      summary: Get a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      summary: Update a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderInput"
      responses:
        "200":
          description: Purchase order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      summary: Delete a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Purchase order deleted
  /purchase-orders/{id}/convert:
    post:
      operationId: convertPurchaseOrderToBill
      tags: [purchase-orders]
      summary: Convert purchase order to bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Bill created from purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
  /sales-leads:
    get:
      operationId: listSalesLeads
      tags: [sales-leads]
      summary: List all sales leads
      responses:
        "200":
          description: List of sales leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesLead"
    post:
      operationId: createSalesLead
      tags: [sales-leads]
      summary: Create a sales lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesLeadInput"
      responses:
        "201":
          description: Sales lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLead"
  /sales-leads/{id}:
    patch:
      operationId: updateSalesLead
      tags: [sales-leads]
      summary: Update a sales lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesLeadInput"
      responses:
        "200":
          description: Sales lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLead"
    delete:
      operationId: deleteSalesLead
      tags: [sales-leads]
      summary: Delete a sales lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Sales lead deleted
components:
  schemas:
...
    LineItem:
      type: object
      properties:
        productId:
          type: integer
          nullable: true
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        taxPercent:
          type: number
          default: 0
        discountPercent:
          type: number
          default: 0
      required:
        - description
        - quantity
        - unitPrice
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        estimateId:
          type: integer
          nullable: true
        quoteId:
          type: integer
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxTotal:
          type: number
        discountTotal:
          type: number
        total:
          type: number
        dueDate:
          type: string
          format: date-time
          nullable: true
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, check, cash]
          nullable: true
        paymentNote:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        isQuickInvoice:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - customerName
        - status
        - lineItems
        - subtotal
        - taxTotal
        - discountTotal
        - total
        - isQuickInvoice
        - createdAt
        - updatedAt
    CreateInvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
        estimateId:
          type: integer
          nullable: true
        quoteId:
          type: integer
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        dueDate:
          type: string
          format: date-time
          nullable: true
        isQuickInvoice:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - customerId
        - lineItems
    UpdateInvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        dueDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        internalNote:
          type: string
          nullable: true
    PayInvoiceInput:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, check, cash]
        paymentNote:
          type: string
          nullable: true
      required:
        - paymentMethod
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        sourceInvoiceId:
          type: integer
          nullable: true
        poSequence:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, sent, received, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxTotal:
          type: number
        total:
          type: number
        notes:
          type: string
          nullable: true
        expectedDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - vendorName
        - status
        - lineItems
        - subtotal
        - taxTotal
        - total
        - createdAt
        - updatedAt
    CreatePurchaseOrderInput:
      type: object
      properties:
        vendorId:
          type: integer
        sourceInvoiceId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, sent, received, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        notes:
          type: string
          nullable: true
        expectedDate:
          type: string
          format: date-time
          nullable: true
      required:
        - vendorId
        - lineItems
    UpdatePurchaseOrderInput:
      type: object
      properties:
        vendorId:
          type: integer
        status:
          type: string
          enum: [draft, sent, received, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        notes:
          type: string
          nullable: true
        expectedDate:
          type: string
          format: date-time
          nullable: true
    Bill:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        purchaseOrderId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, received, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxTotal:
          type: number
        total:
          type: number
        dueDate:
          type: string
          format: date-time
          nullable: true
        paymentMethod:
          type: string
          enum: [bank_transfer, check, cash]
          nullable: true
        paymentNote:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - vendorName
        - status
        - lineItems
        - subtotal
        - taxTotal
        - total
        - createdAt
        - updatedAt
...
    SalesLead:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - createdAt
        - updatedAt
    CreateSalesLeadInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
      required:
        - firstName
        - lastName
    UpdateSalesLeadInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true